-Converting to CMake + began using conan package manager instead of pushing external libs -TextComponent render matrix transforms are now cached which significantly improves performance when multiple of them are visible. Recomputing them every frame used up 10-20% processing time. -UI rework started which will allow to embed canvases into other canvases, procedurally generate their sizes, resize them etc. to produce a user-friendly layout. -Fixed all warnings -float time now converted to Time (which uses a double)

+++ src/src/rendering/Mesh.cpp
@@ -119,8 +119,8 @@ namespace GEE
 			glBufferData(GL_ELEMENT_ARRAY_BUFFER, sizeof(unsigned int) * indices.size(), &(indices)[0], GL_STATIC_DRAW);
 		}
 
-		VertexCount = vertices.size();
-		IndexCount = indices.size();
+		VertexCount = static_cast<unsigned int>(vertices.size());
+		IndexCount = static_cast<unsigned int>(indices.size());
 
 		if (keepVerts)
 		{
